requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
dj-database-url==0.5.0
Django==3.2.8
django-autocomplete-light==3.8.2
django-dal==1.0.0
django-js-asset==1.2.2
django-mptt==0.13.4
django-versatileimagefield==2.2
gunicorn==20.1.0
idna==3.3
Pillow==8.4.0
python-magic==0.4.24
pytz==2021.3
razorpay==1.2.0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==6.2.0
